add bin/koha-print-circ-rule.pl
This is a command line tool which will take a borrower
card number and a item barcode as arguments, and print
the matching line from the circulation rules, plus
a bunch of other information about borrower and
item.

I think that this should be extensible, allowing us
to dump other information which is calculated by Koha
rather than stored in the database.